Add initialization and PID example
